+++ src/views/branchesNode.ts
@@ -9,7 +9,11 @@ export class BranchesNode extends ExplorerNode {
 
         readonly resourceType: ResourceType = 'gitlens:branches';
 
-        constructor(uri: GitUri, protected readonly context: ExtensionContext, protected readonly git: GitService) {
+        constructor(
+            uri: GitUri,
+            protected readonly context: ExtensionContext,
+            protected readonly git: GitService
+        ) {
             super(uri);
         }
 
@@ -18,7 +22,7 @@ export class BranchesNode extends ExplorerNode {
             if (branches === undefined) return [];
 
             branches.sort((a, b) => (a.current ? -1 : 1) - (b.current ? -1 : 1) || a.name.localeCompare(b.name));
-            return [...Iterables.filterMap(branches, b => b.remote ? undefined : new BranchHistoryNode(b, this.uri, this.git.config.gitExplorer.commitFormat, this.context, this.git))];
+            return [...Iterables.filterMap(branches, b => b.remote ? undefined : new BranchHistoryNode(b, this.uri, this.context, this.git))];
         }
 
         async getTreeItem(): Promise<TreeItem> {